[FAB-3160] Move core/config to core/testutil
This will pave the way for avoiding a circular reference once we add
the real support for config-relative pathing as part of FAB-3160
